# api/openapi.yaml
openapi: 3.0.3
info:
  title: BattleBornIntel API
  description: Regional innovation ecosystem intelligence platform for Nevada's startup/venture landscape.
  version: 1.0.0
servers:
  - url: http://localhost:3001/api
    description: Local development
  - url: http://localhost:3001/api/v1
    description: Local development (versioned)
components:
  securitySchemes:
    AdminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: Required for all /admin endpoints.
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
            example:
              error: Too many requests
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
paths:
...
  /funds:
    get:
      tags: [Funds]
      summary: List all funds
      responses:
        "200":
          description: Fund deployment data
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
  /funds/{id}:
    get:
      tags: [Funds]
      summary: Get a single fund by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: "^[a-zA-Z0-9_-]+$"
            maxLength: 20
      responses:
        "200":
          description: Fund detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"

  # ── Graph ──────────────────────────────────────────────────────────────────
...
  /timeline:
    get:
      tags: [Timeline]
      summary: Recent timeline events
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
            maximum: 500
        - name: type
          in: query
          schema:
            type: string
          description: Event type filter
      responses:
        "200":
          description: Timeline events
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
  /timeline/weeks:
    get:
      tags: [Timeline]
      summary: List weeks with event counts
      responses:
        "200":
          description: Week list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
        "500":
          $ref: "#/components/responses/ServerError"
  /timeline/reap:
    get:
      tags: [Timeline]
      summary: MIT REAP framework metrics from timeline events
      parameters:
        - name: since
          in: query
          schema:
            type: string
            format: date
        - name: until
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: REAP metrics
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
  /timeline/week/{date}:
    get:
      tags: [Timeline]
      summary: Events for a specific ISO week
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
          description: Monday of the week (YYYY-MM-DD)
      responses:
        "200":
          description: Week events
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  weekStart:
                    type: string
                    format: date
                  eventCount:
                    type: integer
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"

  # ── Analysis ───────────────────────────────────────────────────────────────
...
  /stakeholder-activities:
    get:
      tags: [Stakeholder Activities]
      summary: Stakeholder activity feed with filters
      parameters:
        - name: location
          in: query
          schema:
            type: string
            default: all
          description: Nevada region slug
        - name: since
          in: query
          schema:
            type: string
            format: date
        - name: until
          in: query
          schema:
            type: string
            format: date
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 500
        - name: type
          in: query
          schema:
            type: string
          description: Activity type (funding, partnership, award, etc.)
        - name: stakeholder_type
          in: query
          schema:
            type: string
            enum: [gov_policy, gov, university, corporate, risk_capital, capital, ecosystem]
      responses:
        "200":
          description: Activity feed with metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  meta:
                    type: object
                    properties:
                      count:
                        type: integer
                      total:
                        type: integer
                      limit:
                        type: integer
                      hasMore:
                        type: boolean
                      filters:
                        type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
  /stakeholder-activities/company/{companyId}:
    get:
      tags: [Stakeholder Activities]
      summary: Activities for a specific company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Company activities
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  meta:
                    type: object
                    properties:
                      count:
                        type: integer
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
  /stakeholder-activities/location/{location}:
    get:
      tags: [Stakeholder Activities]
      summary: Activities by location with date range
      parameters:
        - name: location
          in: path
          required: true
          schema:
            type: string
        - name: since
          in: query
          schema:
            type: string
            format: date
          description: Defaults to 90 days ago
        - name: until
          in: query
          schema:
            type: string
            format: date
          description: Defaults to today
      responses:
        "200":
          description: Location activities
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  meta:
                    type: object
                    properties:
                      count:
                        type: integer
                      location:
                        type: string
                      dateRange:
                        type: object
                        properties:
                          start:
                            type: string
                            format: date
                          end:
                            type: string
                            format: date
        "500":
          $ref: "#/components/responses/ServerError"
  /stakeholder-activities/stats/by-type:
    get:
      tags: [Stakeholder Activities]
      summary: Activity counts grouped by type
      responses:
        "200":
          description: Type counts
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
        "500":
          $ref: "#/components/responses/ServerError"
  /stakeholder-activities/stats/by-location:
    get:
      tags: [Stakeholder Activities]
      summary: Activity counts grouped by location
      responses:
        "200":
          description: Location counts
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
        "500":
          $ref: "#/components/responses/ServerError"

  # ── Constants ──────────────────────────────────────────────────────────────
...